% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{edgeData_keggPathwaygraphs}
\alias{edgeData_keggPathwaygraphs}
\title{utils: utility functions
names in which there is an edge between the given nodes}
\usage{
edgeData_keggPathwaygraphs(edgeFrom, edgeTo, updateKEGG = FALSE)
}
\arguments{
\item{edgeFrom}{A character of node name}

\item{edgeTo}{A character of node name}

\item{updateKEGG}{re-download KEGG data}
}
\value{
A character vector of KEGG pathway names in which
edge (edgeFrom, edgeTo) exists
}
\description{
utils: utility functions
names in which there is an edge between the given nodes
}
\details{
#' Draghici S., et.al.: "A systems biology approach for
pathway level analysis". Genome Research, 17, 2007.
}
\examples{
pathways<-edgeData_keggPathwaygraphs(edgeFrom="hsa:208",
edgeTo="hsa:2309",updateKEGG=FALSE)
class(pathways)
head(pathways)
}
\seealso{
\code{\link{keggPathwayGraphs},\link{keggPathwayNames}}
}
\author{
Azam Peyvandipour and Sorin Draghici
}
