% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggGlobalGraph.R
\name{keggGlobalGraph}
\alias{keggGlobalGraph}
\title{keggGlobalGraph: Construct global graph of KEGG pathways.}
\usage{
keggGlobalGraph(organism = "hsa", updateKEGG = FALSE)
}
\arguments{
\item{organism}{organism code as defined by KEGG}

\item{updateKEGG}{re-download KEGG data}
}
\value{
An object of class \code{\link{graphNEL}} encoding the KEGG
global graph information.
}
\description{
keggGlobalGraph: Construct global graph of KEGG pathways.
}
\examples{
gg<-keggGlobalGraph()

}
\references{
#' Draghici S., et.al.: "A systems biology approach for
pathway level analysis". Genome Research, 17, 2007.
}
\seealso{
\code{\link{keggPathwayGraphs}},
\code{\link{keggPathwayNames}}
}
\author{
Azam Peyvandipour and Sorin Draghici
}
