% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nodeData_keggPathwaygraphs}
\alias{nodeData_keggPathwaygraphs}
\title{edgeData_keggPathwaygraphs: A character vector of
KEGG pathway names in which the given node exists}
\usage{
nodeData_keggPathwaygraphs(node, updateKEGG = FALSE)
}
\arguments{
\item{node}{A character of node name}

\item{updateKEGG}{re-download KEGG data}
}
\value{
A character vector of KEGG pathway names in which the given node exists
}
\description{
edgeData_keggPathwaygraphs: A character vector of
KEGG pathway names in which the given node exists
}
\details{
See cited document for more details.
}
\examples{
pathways<-nodeData_keggPathwaygraphs(node="hsa:208",updateKEGG=FALSE)
}
\seealso{
\code{\link{keggPathwayGraphs},\link{keggPathwayNames}}
}
\author{
Azam Peyvandipour and Sorin Draghici
}
