% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortestPathsGraph.R
\name{shortestPathsGraph}
\alias{shortestPathsGraph}
\title{shortestPathsGraph: Construct a subgraph of KEGG global
graph connecting drug targets and disease-related genes.}
\usage{
shortestPathsGraph(drug_targets, disease_genes, ggraph)
}
\arguments{
\item{drug_targets}{the vector of drug target genes}

\item{disease_genes}{the vector of disease-related genes}

\item{ggraph}{KEGG global graph as object of type \code{\link{graphNEL}}}
}
\value{
An object of class \code{\link{graphNEL}} encoding
the a subgraph of KEGG global graph
}
\description{
shortestPathsGraph: Construct a subgraph of KEGG global
graph connecting drug targets and disease-related genes.
}
\examples{
#obtain KEGG global graph gg
gg<-keggGlobalGraph()
#drug target genes can be obtained from The Comparative
#Toxicogenomics Database (CTD)
drug_targets<-c("hsa:9368","hsa:2322", "hsa:3932", "hsa:4067", "hsa:6714")

#disease-related genes can be obtained from The Comparative
#Toxicogenomics Database (CTD)
disease_genes<-c("hsa:1832" ,"hsa:5073" ,"hsa:5328")

#obtain drug-disease network
mygraph<-shortestPathsGraph(drug_targets,disease_genes,gg)
}
\references{
Draghici S., et.al.: "A systems biology approach for
pathway level analysis".
Genome Research, 17, 2007.
}
\seealso{
\code{\link{keggGlobalGraph}}
}
\author{
Azam Peyvandipour and Sorin Draghici
}
