% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggGlobalGraph.R
\name{updateGlobalGraph}
\alias{updateGlobalGraph}
\title{keggGlobalGraph: Construct global graph of KEGG signaling pathways.}
\usage{
updateGlobalGraph(globalGraph, edgesToremove = list(c("hsa:6657",
  "hsa:79923")))
}
\arguments{
\item{globalGraph}{KEGG global graph as object of type
\code{\link{graphNEL}}}

\item{edgesToremove}{An object of class \code{\link{list}}
encoding edges need to be removed from the graph}
}
\value{
An object of class \code{\link{graphNEL}} encoding the
KEGG global graph information.
}
\description{
keggGlobalGraph: Construct global graph of KEGG signaling pathways.
}
\details{
See cited document for more details.
}
\examples{
library(graph)
gg<-keggGlobalGraph()
class(edges(gg))
length(unlist(edges(gg)))
allEdges<-edges(gg)
allEdges$`hsa:2065`
ggUpdated<-updateGlobalGraph(gg,edgesToremove=
list(c("hsa:2065","hsa:2549"),c("hsa:2065","hsa:25759")))
length(unlist(edges(ggUpdated)))
}
\references{
#' Draghici S., et.al.: "A systems biology approach for
pathway level analysis". Genome Research, 17, 2007.
}
\seealso{
\code{\link{keggPathwayGraphs}}, \code{\link{keggGlobalGraph}}
}
\author{
Azam Peyvandipour and Sorin Draghici
}
